document-genai/requirements.txt
# Unified Document Services Requirements
# Backend dependencies for both classification and extraction

# FastAPI and Server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Database (SQLite)
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# File handling
aiofiles>=23.2.1
PyMuPDF>=1.23.0
Pillow>=10.2.0

# LLM
google-genai>=0.1.0

# Frontend dependencies (Streamlit)
streamlit>=1.31.0
requests>=2.31.0
pandas>=2.1.4

